Day 2 Overhaul
Updates day 2 to a new project.

DS Store issues were creating parsing problems with the project file. Because day 1 and day 3 will be updated as well using this base project, it is more efficient to wipe clean than to sort through all the parsing, or redo all the work on the original project
