--- Meterologize/Meterologize/Model/City.swift
@@ -1,24 +1,0 @@
-//
-//  City.swift
-//  Meterologize
-//
-//  Created by Karl Pfister on 8/19/21.
-//
-
-import Foundation
-
-class City {
-    let name: String
-    let currentTemp: Double
-    let dailyHigh: Int
-    let dailyLow: Int
-    let currentStatus: String
-    
-    init(name: String, currentTemp: Double, dailyHigh: Int, dailyLow: Int, currentStatus: String){
-        self.name = name
-        self.currentTemp = currentTemp
-        self.dailyHigh = dailyHigh
-        self.dailyLow = dailyLow
-        self.currentStatus = currentStatus
-    }
-}
